refactor(test): replace folder location
